turning.py field min&max for bound state at 20ns
data["tb"] and data["tplus"] are the times it takes an uphill electron
starting with W>0 to cross into and out of W-EZ<0, meaning if the field
turns off it's bound.
field_picker() interpolates to find E(t=20), for a variety of times,
including "tb" and "tplus", and saves to "picked.txt". DataFrame has
metadata and "kind" to report what time ("tt", "tb", "tplus") was
interpolated to what value in ns ("tt=20"), and then "W" and "field".
I've made field_picker() pick out tt=20, tt=10, tb=20 and tplus=20. These
are useful things to plot.