Separate functions for DIL and Potential plotting.
I originally had the Potential, including static field, on a subplot
with DIL vs. Field. I now want a plot of just the Potential. I separated
the calculations and plotting for the Potential subplot and the DIL
subplot, and then replaced DIL() with DIL_and_Potential_Plot that takes
the DIL_Plot() and Potential_Plot() and puts them into different
subplots.