main() -> test_convolve()
Running convolution on test data has been broken up and put in it's own
function.
combinations() takes data and a list of keys and returns a dictionary of
the unique values of data[key] and a list of combinations of unique values
from each key.
convolution() takes data DataFrame and E0, Ep, dL, th_LRL and adds the
convolution to data["conv"].
plot_conv() takes the an axes object as well as data and E0... to plot
that convolution.
test_convolve() uses bound_test_data() to build fake data, combinations to
get the run picking criteria, convolution() to build the convolution for
each run, and then plot_convo to plot each run.