main() adds popt, pconv, and fit data to DataFrame
model_func() is used to fit each section of data to
  y = y0 + a*cos(x - x0)
Using a fft works as well, but since this is the method used to analyze
the real data, I'm sticking with this.
main() takes test data from test_convolve() and adds "popt", "pconv",
"fitconv", and "fft" to it, although fft will be removed. It then plots
the fitconv vs conv, showing they are identical. This is as it should be,
there is no "noise" in my convolution like there is in the experiment, so
the result is a perfect sinusoid.