script combines dL +/-1
HarAddTheCos() performs the harmonic addition theorem such that
  a1*cos(x-phi1) + a2*cos(x-phi2) = a*cos(x-phi)
the main script loads params from "params.txt", and finds combos of "E0",
"Ep", and "th_LRL" only, leaving pairs of dL = +/-1. Using the
HarAddTheCos(), a new observation is added to params with dL = np.NaN and
the new y0, a, x0.