main() looks at step-size errors in lab units
main() takes all step-size error observations, converts the values to
lab units.

testing.nb looks at a particular i.c. that has step-size error. It's not
necessarily that the electron comes in on the z axis as previously
thought. Can come from any angle. What is true, is that the
ponderomotive oscillations in energy shift phase by 180 after a
collision. So to avoid an energy shift when skipping over the
singularity, I have to keep in mind what MW phase I stop and restart at.