handles relaunching for step-size errors
Now instead of at handpicked times, the code checks for a step-size
error, figures the relaunch time, relaunches, and loops until a stop
condition. The loop can stop if NDSolve[] reaches tmax successfully, or
if a relaunched orbit is too short to relaunch due to being at such a
small energy.