Parallel 1d model execution
So many changes. Simple_1D_bulk.py runs the simple 1D model from start
to finish, including a randomized toff between 20 and 40 ns, the updated
MW ringdown time of 180 ns, and saving all final energies is a
dataframe, and exporting to "wfinals_1000_u.h5" for example, for a
pulsed field of 100 mV/cm and an uphill electron.

The real speedups are found in numba.jit for many base functions, and
then using lookup tables for turning times with energy resolution of a
bit under 0.1 GHz and field resolution of 1 mV/cm.